+++ src/Illuminate/Foundation/Application.php
@@ -25,7 +25,7 @@ class Application extends Container implements ApplicationContract, HttpKernelIn
      *
      * @var string
      */
-    const VERSION = '5.3.14';
+    const VERSION = '5.3.15';
 
     /**
      * The base path for the Laravel installation.
