+++ src/Illuminate/Notifications/RoutesNotifications.php
@@ -15,7 +15,7 @@ trait RoutesNotifications
      */
     public function notify($instance)
     {
-        app(Dispatcher::class)->send([$this], $instance);
+        app(Dispatcher::class)->send($this, $instance);
     }
 
     /**
