+++ src/Illuminate/Notifications/SendQueuedNotifications.php
@@ -25,14 +25,23 @@ class SendQueuedNotifications implements ShouldQueue
     protected $notification;
 
     /**
+     * All of the channels to send the notification too.
+     *
+     * @var array
+     */
+    protected $channels = null;
+
+    /**
      * Create a new job instance.
      *
      * @param  \Illuminate\Support\Collection  $notifiables
      * @param  \Illuminate\Notifications\Notification  $notification
+     * @param  array  $channels
      * @return void
      */
-    public function __construct($notifiables, $notification)
+    public function __construct($notifiables, $notification, array $channels = null)
     {
+        $this->channels = $channels;
         $this->notifiables = $notifiables;
         $this->notification = $notification;
     }
@@ -45,6 +54,6 @@ class SendQueuedNotifications implements ShouldQueue
      */
     public function handle(ChannelManager $manager)
     {
-        $manager->sendNow($this->notifiables, $this->notification);
+        $manager->sendNow($this->notifiables, $this->notification, $this->channels);
     }
 }
